+++ src/search.rs
@@ -244,7 +244,7 @@ pub fn search(command_params: CommandParams) {
     let mut anis = anis.into_inner().unwrap();
     let learned_ani;
     if !command_params.learned_ani_cmd{
-        learned_ani = regression::use_learned_ani(sketch_params.c, command_params.individual_contig_q, command_params.individual_contig_r, command_params.robust, command_params.median);
+        learned_ani = regression::use_learned_ani(sketch_params.c, command_params.individual_contig_q, command_params.individual_contig_r, command_params.median);
     }
     else{
         learned_ani = command_params.learned_ani;